+++ src/wavelet.c
@@ -37,7 +37,7 @@ static const char help_str[] = "Perform wavelet transform.";
 int main_wavelet(int argc, char* argv[argc])
 {
 	unsigned long flags = 0;
-	long count = 0;
+	int count = 0;
 	long* adims = NULL;
 	const char* in_file = NULL;
 	const char* out_file = NULL;
